+++ containers/vec/test/src/test_vec_59_edge_cases.c
@@ -77,10 +77,10 @@ ERR_59_e test_vec_59_edge_cases(void)
     for (size_t i = 0; i < 4; i++)
     {
         val = malloc(sizeof(i64));
-        *val = i;
+        *val = (i64)i;
         push_back_vec_59(vec, val);
         printf("Assert: i = %lu == %ld = vec->data[%lu]\n", i, *(i64 *)(vec->data[i]), i);
-        assert(i == *(i64 *)(vec->data[i]));
+        assert(i == (size_t)*(i64 *)(vec->data[i]));
     }
     val = malloc(sizeof(i64));
     *val = 59; // Used by later tests
